Extract SiteList and disableUserScroll modifier
