+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/JetpackWelcomeFragment.kt
@@ -5,22 +5,15 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Divider
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -29,10 +22,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.input.nestedscroll.NestedScrollConnection
-import androidx.compose.ui.input.nestedscroll.NestedScrollSource
-import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
@@ -40,13 +29,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Velocity
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.viewmodel.compose.viewModel
-import coil.compose.rememberImagePainter
 import dagger.hilt.android.AndroidEntryPoint
 import org.wordpress.android.R
 import org.wordpress.android.ui.compose.components.ColumnWithFrostedGlassBackground
@@ -58,6 +45,7 @@ import org.wordpress.android.ui.main.jetpack.JetpackWelcomeUiState.Content
 import org.wordpress.android.ui.main.jetpack.JetpackWelcomeUiState.SiteListItem
 import org.wordpress.android.ui.main.jetpack.components.PrimaryButton
 import org.wordpress.android.ui.main.jetpack.components.SecondaryButton
+import org.wordpress.android.ui.main.jetpack.components.SiteList
 
 @AndroidEntryPoint
 class JetpackWelcomeFragment : Fragment() {
@@ -142,8 +130,8 @@ fun ContentState(uiState: Content) {
                 val listState = rememberLazyListState()
                 val blurredListState = rememberLazyListState()
                 SiteList(
-                        uiState.sites,
-                        listState
+                        items = uiState.sites,
+                        listState = listState,
                 )
                 ColumnWithFrostedGlassBackground(
                         blurRadius = 4.dp,
@@ -151,10 +139,11 @@ fun ContentState(uiState: Content) {
                         borderColor = colorResource(R.color.gray_10).copy(alpha = 0.5f),
                         background = { clipModifier, blurModifier ->
                             SiteList(
-                                    uiState.sites,
-                                    blurredListState,
-                                    clipModifier.disableUserScroll(),
-                                    blurModifier,
+                                    items = uiState.sites,
+                                    listState = blurredListState,
+                                    userScrollEnabled = false,
+                                    modifier = clipModifier,
+                                    blurModifier = blurModifier,
                             )
                         }
                 ) {
@@ -177,65 +166,6 @@ fun ContentState(uiState: Content) {
         }
     }
 }
-
-@Composable
-fun SiteList(
-    items: List<SiteListItem>,
-    listState: LazyListState,
-    modifier: Modifier = Modifier,
-    blurModifier: Modifier = Modifier,
-) {
-    LazyColumn(
-            state = listState,
-            modifier = modifier
-                    .background(colorResource(R.color.white))
-                    .padding(horizontal = 30.dp)
-                    .then(blurModifier),
-    ) {
-        items(
-                items = items,
-                key = { it.id },
-        ) { site ->
-            Row(
-                    verticalAlignment = Alignment.CenterVertically,
-            ) {
-                val painter = rememberImagePainter(site.iconUrl) {
-                    placeholder(R.drawable.ic_placeholder_blavatar_grey_lighten_20_40dp)
-                    crossfade(true)
-                }
-                Image(
-                        painter = painter,
-                        contentDescription = null,
-                        modifier = Modifier
-                                .padding(vertical = 15.dp)
-                                .padding(end = 20.dp)
-                                .size(60.dp)
-                                .clip(RoundedCornerShape(3.dp))
-                )
-                Column {
-                    Text(
-                            text = site.name,
-                            fontWeight = FontWeight.SemiBold,
-                            fontSize = 17.sp,
-                    )
-                    Text(
-                            text = site.url,
-                            fontSize = FontSize.Large.value,
-                            color = colorResource(R.color.gray_40)
-                    )
-                }
-            }
-            Divider(color = colorResource(R.color.gray_10).copy(alpha = 0.5f))
-        }
-    }
-}
-
-private fun Modifier.disableUserScroll() = nestedScroll(
-        connection = object : NestedScrollConnection {
-            override fun onPreScroll(available: Offset, source: NestedScrollSource) = available.copy(x = 0f)
-            override suspend fun onPreFling(available: Velocity) = available.copy(x = 0f)
-        }
-)
 
 fun previewSiteListItems(): List<SiteListItem> {
     val list = mutableListOf<SiteListItem>()
